Fix manifest warnings (#9)
* Remove superfluous file from MANIFEST

It looks like this file had been around at the time of MANIFEST
creation, however wasn't part of the repository.  This silences a
warning at Makefile build time that the file is missing from the dist.

* Fix README filename in MANIFEST

The README has been converted to POD and as such now has a `.pod`
extension.  This change brings the MANIFEST up to date.