Fix POD warnings (#15)
* Move run() method POD to main docs section

`podchecker` noticed that the `=head2` was missing a preceding `=head1`;
upon closer inspection it seemed that the POD for the `run()` method
should have been part of the main documentation section which is at the
end of the file.  Moving the docs for `run()` to the main section fixes
the podchecker warning and puts the docs for this file all in the one
location.

* Make SWISH_DOC_PROP_MAP heading into an item

Since, in the given context, it's an item in list.  This also fixes a
podchecker warning about an empty paragraph under the heading.

* Add missing docs for SWISH-related methods

Some of the docs were missing for SWISH-related methods.  The text used
here for each of the items is taken from the SWISH-e source code.

* Add SYNOPSIS for Dezi::Test::Doc

* Document methods in Dezi::Test::InvIndex

* Document next() method of Dezi::Test::Results