Replace maildir file with DATA var in mail tests
This removes the need to have a non-portable file in the distribution
and repository (see the `portable_filenames` CPANTS issue) and makes the
mail tests more self contained.  Also, this can improve packaging since
the `.INBOX` directory no longer needs to exist as part of the package:
this directory and the associated mbox file can simply be created as
part of the test and then removed afterwards.